Move problem generator inside the motion planner
