+++ tesseract_motion_planners/examples/raster_example.cpp
@@ -38,11 +38,9 @@ TESSERACT_COMMON_IGNORE_WARNINGS_POP
 #include <tesseract_motion_planners/descartes/descartes_collision.h>
 #include <tesseract_motion_planners/descartes/descartes_motion_planner.h>
 #include <tesseract_motion_planners/descartes/descartes_utils.h>
-#include <tesseract_motion_planners/descartes/problem_generators/default_problem_generator.h>
 #include <tesseract_motion_planners/descartes/profile/descartes_default_plan_profile.h>
 
 #include <tesseract_motion_planners/trajopt/trajopt_motion_planner.h>
-#include <tesseract_motion_planners/trajopt/problem_generators/default_problem_generator.h>
 #include <tesseract_motion_planners/trajopt/profile/trajopt_default_plan_profile.h>
 #include <tesseract_motion_planners/trajopt/profile/trajopt_default_composite_profile.h>
 
@@ -84,229 +82,235 @@ std::string locateResource(const std::string& url)
 
 int main(int /*argc*/, char** /*argv*/)
 {
-  // Setup
-  auto locator = std::make_shared<tesseract_common::SimpleResourceLocator>(locateResource);
-  auto env = std::make_shared<tesseract_environment::Environment>();
-  tesseract_common::fs::path urdf_path(std::string(TESSERACT_SUPPORT_DIR) + "/urdf/abb_irb2400.urdf");
-  tesseract_common::fs::path srdf_path(std::string(TESSERACT_SUPPORT_DIR) + "/urdf/abb_irb2400.srdf");
-  env->init(urdf_path, srdf_path, locator);
-
-  // Dynamically load ignition visualizer if exist
-  tesseract_visualization::VisualizationLoader loader;
-  auto plotter = loader.get();
-
-  if (plotter != nullptr)
+  try
   {
-    plotter->waitForConnection();
-    plotter->plotEnvironment(*env);
+    // Setup
+    auto locator = std::make_shared<tesseract_common::SimpleResourceLocator>(locateResource);
+    auto env = std::make_shared<tesseract_environment::Environment>();
+    tesseract_common::fs::path urdf_path(std::string(TESSERACT_SUPPORT_DIR) + "/urdf/abb_irb2400.urdf");
+    tesseract_common::fs::path srdf_path(std::string(TESSERACT_SUPPORT_DIR) + "/urdf/abb_irb2400.srdf");
+    env->init(urdf_path, srdf_path, locator);
+
+    // Dynamically load ignition visualizer if exist
+    tesseract_visualization::VisualizationLoader loader;
+    auto plotter = loader.get();
+
+    if (plotter != nullptr)
+    {
+      plotter->waitForConnection();
+      plotter->plotEnvironment(*env);
+    }
+
+    ManipulatorInfo manip;
+    manip.tcp_frame = "tool0";
+    manip.manipulator = "manipulator";
+    manip.manipulator_ik_solver = "OPWInvKin";
+
+    auto kin_group = env->getKinematicGroup(manip.manipulator, manip.manipulator_ik_solver);
+    auto cur_state = env->getState();
+
+    CompositeInstruction program("raster_program", CompositeInstructionOrder::ORDERED, manip);
+
+    // Start Joint Position for the program
+    Waypoint wp0 = StateWaypoint(kin_group->getJointNames(), Eigen::VectorXd::Zero(6));
+    PlanInstruction start_instruction(wp0, PlanInstructionType::START);
+    program.setStartInstruction(start_instruction);
+
+    // Define raster poses
+    Waypoint wp1 = CartesianWaypoint(Eigen::Isometry3d::Identity() * Eigen::Translation3d(0.8, -0.3, 0.8) *
+                                     Eigen::Quaterniond(0, 0, -1.0, 0));
+    Waypoint wp2 = CartesianWaypoint(Eigen::Isometry3d::Identity() * Eigen::Translation3d(0.8, -0.2, 0.8) *
+                                     Eigen::Quaterniond(0, 0, -1.0, 0));
+    Waypoint wp3 = CartesianWaypoint(Eigen::Isometry3d::Identity() * Eigen::Translation3d(0.8, -0.1, 0.8) *
+                                     Eigen::Quaterniond(0, 0, -1.0, 0));
+    Waypoint wp4 = CartesianWaypoint(Eigen::Isometry3d::Identity() * Eigen::Translation3d(0.8, 0.0, 0.8) *
+                                     Eigen::Quaterniond(0, 0, -1.0, 0));
+    Waypoint wp5 = CartesianWaypoint(Eigen::Isometry3d::Identity() * Eigen::Translation3d(0.8, 0.1, 0.8) *
+                                     Eigen::Quaterniond(0, 0, -1.0, 0));
+    Waypoint wp6 = CartesianWaypoint(Eigen::Isometry3d::Identity() * Eigen::Translation3d(0.8, 0.2, 0.8) *
+                                     Eigen::Quaterniond(0, 0, -1.0, 0));
+    Waypoint wp7 = CartesianWaypoint(Eigen::Isometry3d::Identity() * Eigen::Translation3d(0.8, 0.3, 0.8) *
+                                     Eigen::Quaterniond(0, 0, -1.0, 0));
+
+    // Define raster move instruction
+    PlanInstruction plan_c0(wp2, PlanInstructionType::LINEAR, "RASTER");
+    PlanInstruction plan_c1(wp3, PlanInstructionType::LINEAR, "RASTER");
+    PlanInstruction plan_c2(wp4, PlanInstructionType::LINEAR, "RASTER");
+    PlanInstruction plan_c3(wp5, PlanInstructionType::LINEAR, "RASTER");
+    PlanInstruction plan_c4(wp6, PlanInstructionType::LINEAR, "RASTER");
+    PlanInstruction plan_c5(wp7, PlanInstructionType::LINEAR, "RASTER");
+
+    PlanInstruction plan_f0(wp1, PlanInstructionType::FREESPACE, "freespace_profile");
+    plan_f0.setDescription("from_start_plan");
+    CompositeInstruction from_start;
+    from_start.setDescription("from_start");
+    from_start.push_back(plan_f0);
+    program.push_back(from_start);
+
+    {
+      CompositeInstruction raster_segment;
+      raster_segment.setDescription("raster_segment");
+      raster_segment.push_back(plan_c0);
+      raster_segment.push_back(plan_c1);
+      raster_segment.push_back(plan_c2);
+      raster_segment.push_back(plan_c3);
+      raster_segment.push_back(plan_c4);
+      raster_segment.push_back(plan_c5);
+      program.push_back(raster_segment);
+    }
+
+    {
+      PlanInstruction plan_f1(wp1, PlanInstructionType::FREESPACE, "freespace_profile");
+      plan_f1.setDescription("transition_from_end_plan");
+      CompositeInstruction transition_from_end;
+      transition_from_end.setDescription("transition_from_end");
+      transition_from_end.push_back(plan_f1);
+      CompositeInstruction transition_from_start;
+      transition_from_start.setDescription("transition_from_start");
+      transition_from_start.push_back(plan_f1);
+
+      CompositeInstruction transitions("DEFAULT", CompositeInstructionOrder::UNORDERED);
+      transitions.setDescription("transitions");
+      transitions.push_back(transition_from_start);
+      transitions.push_back(transition_from_end);
+      program.push_back(transitions);
+    }
+
+    {
+      CompositeInstruction raster_segment;
+      raster_segment.setDescription("raster_segment");
+      raster_segment.push_back(plan_c0);
+      raster_segment.push_back(plan_c1);
+      raster_segment.push_back(plan_c2);
+      raster_segment.push_back(plan_c3);
+      raster_segment.push_back(plan_c4);
+      raster_segment.push_back(plan_c5);
+      program.push_back(raster_segment);
+    }
+
+    {
+      PlanInstruction plan_f1(wp1, PlanInstructionType::FREESPACE, "freespace_profile");
+      plan_f1.setDescription("transition_from_end_plan");
+      CompositeInstruction transition_from_end;
+      transition_from_end.setDescription("transition_from_end");
+      transition_from_end.push_back(plan_f1);
+      CompositeInstruction transition_from_start;
+      transition_from_start.setDescription("transition_from_start");
+      transition_from_start.push_back(plan_f1);
+
+      CompositeInstruction transitions("DEFAULT", CompositeInstructionOrder::UNORDERED);
+      transitions.setDescription("transitions");
+      transitions.push_back(transition_from_start);
+      transitions.push_back(transition_from_end);
+      program.push_back(transitions);
+    }
+
+    {
+      CompositeInstruction raster_segment;
+      raster_segment.setDescription("raster_segment");
+      raster_segment.push_back(plan_c0);
+      raster_segment.push_back(plan_c1);
+      raster_segment.push_back(plan_c2);
+      raster_segment.push_back(plan_c3);
+      raster_segment.push_back(plan_c4);
+      raster_segment.push_back(plan_c5);
+      program.push_back(raster_segment);
+    }
+
+    PlanInstruction plan_f2(wp1, PlanInstructionType::FREESPACE, "freespace_profile");
+    plan_f2.setDescription("to_end_plan");
+    CompositeInstruction to_end;
+    to_end.setDescription("to_end");
+    to_end.push_back(plan_f2);
+    program.push_back(to_end);
+
+    // Plot Program
+    auto state_solver = env->getStateSolver();
+    if (plotter)
+    {
+    }
+
+    // Create Profiles
+    auto descartes_plan_profile = std::make_shared<DescartesDefaultPlanProfileD>();
+    auto trajopt_plan_profile = std::make_shared<TrajOptDefaultPlanProfile>();
+    auto trajopt_composite_profile = std::make_shared<TrajOptDefaultCompositeProfile>();
+
+    // Create a seed
+    CompositeInstruction seed = generateSeed(program, cur_state, env);
+
+    // Profile Dictionary
+    auto profiles = std::make_shared<ProfileDictionary>();
+    profiles->addProfile<DescartesPlanProfile<double>>("DEFAULT", descartes_plan_profile);
+    profiles->addProfile<TrajOptPlanProfile>("DEFAULT", trajopt_plan_profile);
+    profiles->addProfile<TrajOptCompositeProfile>("DEFAULT", trajopt_composite_profile);
+
+    // Create Planning Request
+    PlannerRequest request;
+    request.seed = seed;
+    request.instructions = program;
+    request.env = env;
+    request.env_state = cur_state;
+    request.profiles = profiles;
+
+    // Solve Descartes Plan
+    PlannerResponse descartes_response;
+    DescartesMotionPlannerD descartes_planner;
+    auto descartes_status = descartes_planner.solve(request, descartes_response);
+    assert(descartes_status);
+
+    // Plot Descartes Trajectory
+    if (plotter)
+    {
+      plotter->waitForInput();
+      plotter->plotTrajectory(toJointTrajectory(descartes_response.results), *state_solver);
+    }
+
+    // Update Seed
+    request.seed = descartes_response.results;
+
+    // Solve TrajOpt Plan
+    PlannerResponse trajopt_response;
+    TrajOptMotionPlanner trajopt_planner;
+    auto trajopt_status = trajopt_planner.solve(request, trajopt_response);
+    assert(trajopt_status);
+
+    if (plotter)
+    {
+      plotter->waitForInput();
+      plotter->plotTrajectory(toJointTrajectory(trajopt_response.results), *state_solver);
+    }
+
+    //  // *************************************
+    //  // Create Motion Plan for home to raster
+    //  // *************************************
+
+    //  // Get raster solution last position
+    //  JointWaypoint last_position;  // TODO: Get last move instruction position
+
+    //  // Update last plan instruction
+    //  p1.setEndWaypoint(last_position);
+
+    //  // Create a seed
+    //  CompositeInstruction p1_seed = generateSeed(p1, cur_state, fwd_kin, inv_kin);
+
+    //  // Use planners
+
+    //  // *************************************
+    //  // Create Motion Plan for raster to home
+    //  // *************************************
+
+    //  // Get raster solution last position
+    //  JointWaypoint first_position;  // TODO: Get first waypoint
+
+    //  // Update end waypoint
+    //  p1.setStartWaypoint(first_position);
+
+    //  // Create a seed
+    //  CompositeInstruction p2_seed = generateSeed(p2, cur_state, fwd_kin, inv_kin);
   }
-
-  ManipulatorInfo manip;
-  manip.tcp_frame = "tool0";
-  manip.manipulator = "manipulator";
-  manip.manipulator_ik_solver = "OPWInvKin";
-
-  auto kin_group = env->getKinematicGroup(manip.manipulator, manip.manipulator_ik_solver);
-  auto cur_state = env->getState();
-
-  CompositeInstruction program("raster_program", CompositeInstructionOrder::ORDERED, manip);
-
-  // Start Joint Position for the program
-  Waypoint wp0 = StateWaypoint(kin_group->getJointNames(), Eigen::VectorXd::Zero(6));
-  PlanInstruction start_instruction(wp0, PlanInstructionType::START);
-  program.setStartInstruction(start_instruction);
-
-  // Define raster poses
-  Waypoint wp1 = CartesianWaypoint(Eigen::Isometry3d::Identity() * Eigen::Translation3d(0.8, -0.3, 0.8) *
-                                   Eigen::Quaterniond(0, 0, -1.0, 0));
-  Waypoint wp2 = CartesianWaypoint(Eigen::Isometry3d::Identity() * Eigen::Translation3d(0.8, -0.2, 0.8) *
-                                   Eigen::Quaterniond(0, 0, -1.0, 0));
-  Waypoint wp3 = CartesianWaypoint(Eigen::Isometry3d::Identity() * Eigen::Translation3d(0.8, -0.1, 0.8) *
-                                   Eigen::Quaterniond(0, 0, -1.0, 0));
-  Waypoint wp4 = CartesianWaypoint(Eigen::Isometry3d::Identity() * Eigen::Translation3d(0.8, 0.0, 0.8) *
-                                   Eigen::Quaterniond(0, 0, -1.0, 0));
-  Waypoint wp5 = CartesianWaypoint(Eigen::Isometry3d::Identity() * Eigen::Translation3d(0.8, 0.1, 0.8) *
-                                   Eigen::Quaterniond(0, 0, -1.0, 0));
-  Waypoint wp6 = CartesianWaypoint(Eigen::Isometry3d::Identity() * Eigen::Translation3d(0.8, 0.2, 0.8) *
-                                   Eigen::Quaterniond(0, 0, -1.0, 0));
-  Waypoint wp7 = CartesianWaypoint(Eigen::Isometry3d::Identity() * Eigen::Translation3d(0.8, 0.3, 0.8) *
-                                   Eigen::Quaterniond(0, 0, -1.0, 0));
-
-  // Define raster move instruction
-  PlanInstruction plan_c0(wp2, PlanInstructionType::LINEAR, "RASTER");
-  PlanInstruction plan_c1(wp3, PlanInstructionType::LINEAR, "RASTER");
-  PlanInstruction plan_c2(wp4, PlanInstructionType::LINEAR, "RASTER");
-  PlanInstruction plan_c3(wp5, PlanInstructionType::LINEAR, "RASTER");
-  PlanInstruction plan_c4(wp6, PlanInstructionType::LINEAR, "RASTER");
-  PlanInstruction plan_c5(wp7, PlanInstructionType::LINEAR, "RASTER");
-
-  PlanInstruction plan_f0(wp1, PlanInstructionType::FREESPACE, "freespace_profile");
-  plan_f0.setDescription("from_start_plan");
-  CompositeInstruction from_start;
-  from_start.setDescription("from_start");
-  from_start.push_back(plan_f0);
-  program.push_back(from_start);
-
+  catch (const std::exception& e)
   {
-    CompositeInstruction raster_segment;
-    raster_segment.setDescription("raster_segment");
-    raster_segment.push_back(plan_c0);
-    raster_segment.push_back(plan_c1);
-    raster_segment.push_back(plan_c2);
-    raster_segment.push_back(plan_c3);
-    raster_segment.push_back(plan_c4);
-    raster_segment.push_back(plan_c5);
-    program.push_back(raster_segment);
+    CONSOLE_BRIDGE_logError("Example failed with message: %s", e.what());
+    return -1;
   }
-
-  {
-    PlanInstruction plan_f1(wp1, PlanInstructionType::FREESPACE, "freespace_profile");
-    plan_f1.setDescription("transition_from_end_plan");
-    CompositeInstruction transition_from_end;
-    transition_from_end.setDescription("transition_from_end");
-    transition_from_end.push_back(plan_f1);
-    CompositeInstruction transition_from_start;
-    transition_from_start.setDescription("transition_from_start");
-    transition_from_start.push_back(plan_f1);
-
-    CompositeInstruction transitions("DEFAULT", CompositeInstructionOrder::UNORDERED);
-    transitions.setDescription("transitions");
-    transitions.push_back(transition_from_start);
-    transitions.push_back(transition_from_end);
-    program.push_back(transitions);
-  }
-
-  {
-    CompositeInstruction raster_segment;
-    raster_segment.setDescription("raster_segment");
-    raster_segment.push_back(plan_c0);
-    raster_segment.push_back(plan_c1);
-    raster_segment.push_back(plan_c2);
-    raster_segment.push_back(plan_c3);
-    raster_segment.push_back(plan_c4);
-    raster_segment.push_back(plan_c5);
-    program.push_back(raster_segment);
-  }
-
-  {
-    PlanInstruction plan_f1(wp1, PlanInstructionType::FREESPACE, "freespace_profile");
-    plan_f1.setDescription("transition_from_end_plan");
-    CompositeInstruction transition_from_end;
-    transition_from_end.setDescription("transition_from_end");
-    transition_from_end.push_back(plan_f1);
-    CompositeInstruction transition_from_start;
-    transition_from_start.setDescription("transition_from_start");
-    transition_from_start.push_back(plan_f1);
-
-    CompositeInstruction transitions("DEFAULT", CompositeInstructionOrder::UNORDERED);
-    transitions.setDescription("transitions");
-    transitions.push_back(transition_from_start);
-    transitions.push_back(transition_from_end);
-    program.push_back(transitions);
-  }
-
-  {
-    CompositeInstruction raster_segment;
-    raster_segment.setDescription("raster_segment");
-    raster_segment.push_back(plan_c0);
-    raster_segment.push_back(plan_c1);
-    raster_segment.push_back(plan_c2);
-    raster_segment.push_back(plan_c3);
-    raster_segment.push_back(plan_c4);
-    raster_segment.push_back(plan_c5);
-    program.push_back(raster_segment);
-  }
-
-  PlanInstruction plan_f2(wp1, PlanInstructionType::FREESPACE, "freespace_profile");
-  plan_f2.setDescription("to_end_plan");
-  CompositeInstruction to_end;
-  to_end.setDescription("to_end");
-  to_end.push_back(plan_f2);
-  program.push_back(to_end);
-
-  // Plot Program
-  auto state_solver = env->getStateSolver();
-  if (plotter)
-  {
-  }
-
-  // Create Profiles
-  auto descartes_plan_profile = std::make_shared<DescartesDefaultPlanProfileD>();
-  auto trajopt_plan_profile = std::make_shared<TrajOptDefaultPlanProfile>();
-  auto trajopt_composite_profile = std::make_shared<TrajOptDefaultCompositeProfile>();
-
-  // Create a seed
-  CompositeInstruction seed = generateSeed(program, cur_state, env);
-
-  // Profile Dictionary
-  auto profiles = std::make_shared<ProfileDictionary>();
-  profiles->addProfile<DescartesPlanProfile<double>>("DEFAULT", descartes_plan_profile);
-  profiles->addProfile<TrajOptPlanProfile>("DEFAULT", trajopt_plan_profile);
-  profiles->addProfile<TrajOptCompositeProfile>("DEFAULT", trajopt_composite_profile);
-
-  // Create Planning Request
-  PlannerRequest request;
-  request.seed = seed;
-  request.instructions = program;
-  request.env = env;
-  request.env_state = cur_state;
-  request.profiles = profiles;
-
-  // Solve Descartes Plan
-  PlannerResponse descartes_response;
-  DescartesMotionPlannerD descartes_planner;
-  descartes_planner.problem_generator = tesseract_planning::DefaultDescartesProblemGenerator<double>;
-  auto descartes_status = descartes_planner.solve(request, descartes_response);
-  assert(descartes_status);
-
-  // Plot Descartes Trajectory
-  if (plotter)
-  {
-    plotter->waitForInput();
-    plotter->plotTrajectory(toJointTrajectory(descartes_response.results), *state_solver);
-  }
-
-  // Update Seed
-  request.seed = descartes_response.results;
-
-  // Solve TrajOpt Plan
-  PlannerResponse trajopt_response;
-  TrajOptMotionPlanner trajopt_planner;
-  trajopt_planner.problem_generator = tesseract_planning::DefaultTrajoptProblemGenerator;
-  auto trajopt_status = trajopt_planner.solve(request, trajopt_response);
-  assert(trajopt_status);
-
-  if (plotter)
-  {
-    plotter->waitForInput();
-    plotter->plotTrajectory(toJointTrajectory(trajopt_response.results), *state_solver);
-  }
-
-  //  // *************************************
-  //  // Create Motion Plan for home to raster
-  //  // *************************************
-
-  //  // Get raster solution last position
-  //  JointWaypoint last_position;  // TODO: Get last move instruction position
-
-  //  // Update last plan instruction
-  //  p1.setEndWaypoint(last_position);
-
-  //  // Create a seed
-  //  CompositeInstruction p1_seed = generateSeed(p1, cur_state, fwd_kin, inv_kin);
-
-  //  // Use planners
-
-  //  // *************************************
-  //  // Create Motion Plan for raster to home
-  //  // *************************************
-
-  //  // Get raster solution last position
-  //  JointWaypoint first_position;  // TODO: Get first waypoint
-
-  //  // Update end waypoint
-  //  p1.setStartWaypoint(first_position);
-
-  //  // Create a seed
-  //  CompositeInstruction p2_seed = generateSeed(p2, cur_state, fwd_kin, inv_kin);
 }
