+++ tesseract_motion_planners/trajopt/CMakeLists.txt
@@ -10,7 +10,6 @@ add_library(
   src/profile/trajopt_default_plan_profile.cpp
   src/profile/trajopt_default_composite_profile.cpp
   src/profile/trajopt_default_solver_profile.cpp
-  src/problem_generators/default_problem_generator.cpp
   src/serialize.cpp
   src/deserialize.cpp)
 target_link_libraries(
