+++ tesseract_process_managers/src/taskflow_generators/cartesian_taskflow.cpp
@@ -41,11 +41,9 @@ TESSERACT_COMMON_IGNORE_WARNINGS_POP
 #include <tesseract_motion_planners/simple/profile/simple_planner_profile.h>
 
 #include <tesseract_motion_planners/descartes/descartes_motion_planner.h>
-#include <tesseract_motion_planners/descartes/problem_generators/default_problem_generator.h>
 #include <tesseract_motion_planners/descartes/profile/descartes_profile.h>
 
 #include <tesseract_motion_planners/trajopt/trajopt_motion_planner.h>
-#include <tesseract_motion_planners/trajopt/problem_generators/default_problem_generator.h>
 #include <tesseract_motion_planners/trajopt/profile/trajopt_profile.h>
 
 using namespace tesseract_planning;
@@ -107,14 +105,12 @@ TaskflowContainer CartesianTaskflow::generateTaskflow(TaskInput input, TaskflowV
 
   // Setup Descartes
   auto descartes_planner = std::make_shared<DescartesMotionPlanner<float>>();
-  descartes_planner->problem_generator = &DefaultDescartesProblemGenerator<float>;
   auto descartes_generator = std::make_unique<MotionPlannerTaskGenerator>(descartes_planner);
   descartes_generator->assignConditionalTask(input, descartes_task);
   container.generators.push_back(std::move(descartes_generator));
 
   // Setup TrajOpt
   auto trajopt_planner = std::make_shared<TrajOptMotionPlanner>();
-  trajopt_planner->problem_generator = &DefaultTrajoptProblemGenerator;
   TaskGenerator::UPtr trajopt_generator = std::make_unique<MotionPlannerTaskGenerator>(trajopt_planner);
   trajopt_generator->assignConditionalTask(input, trajopt_task);
   container.generators.push_back(std::move(trajopt_generator));
